# docs/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autograder API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sandbox/courses/{course_id}/assignments": {
      "get": {
        "tags": [
          "sandbox"
        ],
        "summary": "List Assignments",
        "operationId": "list_assignments_sandbox_courses__course_id__assignments_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxAssignmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandbox/courses/{course_id}/assignments/{assignment_id}": {
      "get": {
        "tags": [
          "sandbox"
        ],
        "summary": "Get Assignment",
        "operationId": "get_assignment_sandbox_courses__course_id__assignments__assignment_id__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxAssignmentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandbox/courses/{course_id}/assignments/{assignment_id}/runs": {
      "post": {
        "tags": [
          "sandbox"
        ],
        "summary": "Create Run",
        "description": "Create a new sandbox run for a submission bundle, with optional stdin input.",
        "operationId": "create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxRunCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandbox/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "sandbox"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run_sandbox_runs__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxCancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandbox/runs/{run_id}/result": {
      "get": {
        "tags": [
          "sandbox"
        ],
        "summary": "Get Result",
        "operationId": "get_result_sandbox_runs__run_id__result_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxRunResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandbox/runs/{run_id}/ai-feedback": {
      "post": {
        "tags": [
          "sandbox"
        ],
        "summary": "Generate Ai Feedback Endpoint",
        "operationId": "generate_ai_feedback_endpoint_sandbox_runs__run_id__ai_feedback_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxAiFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandbox/runs/{run_id}/ai-feedback/stream": {
      "post": {
        "tags": [
          "sandbox"
        ],
        "summary": "Generate Ai Feedback Stream Endpoint",
        "operationId": "generate_ai_feedback_stream_endpoint_sandbox_runs__run_id__ai_feedback_stream_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-sandbox-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Sandbox-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post": {
        "properties": {
          "bundle": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Bundle"
          },
          "stdin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stdin"
          }
        },
        "type": "object",
        "required": [
          "bundle"
        ],
        "title": "Body_create_run_sandbox_courses__course_id__assignments__assignment_id__runs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "queue",
              "run",
              "complete",
              "failure"
            ],
            "title": "State"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "eta_band": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "under_1_min",
                  "1_to_3_min",
                  "3_to_5_min",
                  "over_5_min"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Band"
          },
          "counters": {
            "$ref": "#/components/schemas/RunCounters"
          },
          "backpressure": {
            "$ref": "#/components/schemas/QueueBackpressure"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "state",
          "counters",
          "backpressure"
        ],
        "title": "RunStatusResponse"
      },
...
      "SandboxAiFeedbackResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "ai_feedback": {
            "type": "string",
            "title": "Ai Feedback"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "ai_feedback",
          "model"
        ],
        "title": "SandboxAiFeedbackResponse"
      },
      "SandboxAssignmentDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sandbox_enabled": {
            "type": "boolean",
            "title": "Sandbox Enabled"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "max_score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Score"
          },
          "upload_quota": {
            "$ref": "#/components/schemas/UploadQuota"
          },
          "module_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "accepted_bundle_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Bundle Types"
          },
          "max_upload_bytes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Upload Bytes"
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/SandboxConstraint"
            },
            "type": "array",
            "title": "Constraints"
          },
          "allowed_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Concepts"
          },
          "rubric": {
            "items": {
              "$ref": "#/components/schemas/SandboxRubricItem"
            },
            "type": "array",
            "title": "Rubric"
          },
          "rubric_groups": {
            "items": {
              "$ref": "#/components/schemas/SandboxRubricGroup"
            },
            "type": "array",
            "title": "Rubric Groups"
          },
          "completion_requirements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Completion Requirements"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "title",
          "sandbox_enabled",
          "language",
          "max_score",
          "upload_quota",
          "description",
          "accepted_bundle_types",
          "max_upload_bytes",
          "constraints",
          "rubric"
        ],
        "title": "SandboxAssignmentDetail"
      },
      "SandboxAssignmentListResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/SandboxAssignmentSummary"
            },
            "type": "array",
            "title": "Assignments"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "assignments"
        ],
        "title": "SandboxAssignmentListResponse"
      },
      "SandboxAssignmentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sandbox_enabled": {
            "type": "boolean",
            "title": "Sandbox Enabled"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "max_score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Score"
          },
          "upload_quota": {
            "$ref": "#/components/schemas/UploadQuota"
          },
          "module_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "title",
          "sandbox_enabled",
          "language",
          "max_score",
          "upload_quota"
        ],
        "title": "SandboxAssignmentSummary"
      },
      "SandboxCancelResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "type": "string",
            "const": "failure",
            "title": "State"
          },
          "counters": {
            "$ref": "#/components/schemas/RunCounters"
          },
          "backpressure": {
            "$ref": "#/components/schemas/QueueBackpressure"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "state",
          "counters",
          "backpressure",
          "message"
        ],
        "title": "SandboxCancelResponse"
      },
      "SandboxConstraint": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "SandboxConstraint"
      },
...
      "SandboxRubricGroup": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "item_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Item Keys"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "item_keys"
        ],
        "title": "SandboxRubricGroup"
      },
      "SandboxRubricItem": {
        "properties": {
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "points": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Points"
          },
          "extra_credit": {
            "type": "boolean",
            "title": "Extra Credit",
            "default": false
          },
          "pytest_marker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pytest Marker"
          },
          "item_type": {
            "type": "string",
            "enum": [
              "pytest",
              "manual"
            ],
            "title": "Item Type",
            "default": "pytest"
          },
          "rubric_group_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Group Key"
          },
          "inputs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputs"
          },
          "outputs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs"
          }
        },
        "type": "object",
        "required": [
          "label",
          "points"
        ],
        "title": "SandboxRubricItem"
      },
      "SandboxRunCreateResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "sandbox_session": {
            "type": "string",
            "title": "Sandbox Session"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "result_url": {
            "type": "string",
            "title": "Result Url"
          },
          "upload_quota": {
            "$ref": "#/components/schemas/UploadQuota"
          },
          "initial_status": {
            "$ref": "#/components/schemas/RunStatusResponse"
          },
          "file_preview": {
            "$ref": "#/components/schemas/FilePreviewMetadata"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "sandbox_session",
          "status_url",
          "result_url",
          "upload_quota",
          "initial_status",
          "file_preview"
        ],
        "title": "SandboxRunCreateResponse"
      },
      "SandboxRunResultResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "complete",
              "failure"
            ],
            "title": "State"
          },
          "projected_score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Projected Score"
          },
          "max_score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Score"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/SandboxWarning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "test_summaries": {
            "items": {
              "$ref": "#/components/schemas/TestSummary"
            },
            "type": "array",
            "title": "Test Summaries"
          },
          "rubric_groups": {
            "items": {
              "$ref": "#/components/schemas/RubricGroupResultResponse"
            },
            "type": "array",
            "title": "Rubric Groups"
          },
          "sanitized_feedback": {
            "type": "string",
            "title": "Sanitized Feedback"
          },
          "ai_feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Feedback"
          },
          "file_preview": {
            "$ref": "#/components/schemas/FilePreviewMetadata"
          },
          "retention_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Notice"
          },
          "raw_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Output"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "state",
          "projected_score",
          "max_score",
          "warnings",
          "test_summaries",
          "sanitized_feedback",
          "file_preview"
        ],
        "title": "SandboxRunResultResponse"
      },
...
      "UploadQuota": {
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "default": 5
          },
          "window_seconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Window Seconds",
            "default": 3600
          },
          "remaining": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Remaining"
          },
          "reset_at": {
            "type": "string",
            "title": "Reset At"
          }
        },
        "type": "object",
        "required": [
          "remaining",
          "reset_at"
        ],
        "title": "UploadQuota"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}